requirements.txt
flask
google-genai
Pillow
gunicorn
python-dotenv
